+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 2019-08-02 v2.6.0-SNAPSHOT 
 * Fixed security vulnerability with jackson databind
 * Changed response status on partial success of batch save/update - return 201/200 if at least one of the records was saved/updated, 500 if none of the records was saved/updated
+* Refactored DAOs to return saved/updated entities
 * Added suppress from discovery endpoint for change records value
 
 | METHOD |             URL                                       | DESCRIPTION                                      |